refactoring for the get themes route to work in index.js
